feat(spx-gui): use an angle picker in the code editor to control the sprite direction (#1956)
